+++ CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
       - adding generic file filter extension (.*) since some desktop environments filters are case sensitive (https://github.com/vinifmor/bauh/issues/185)[#185]
       - generating a default **.desktop** file based on the installation form for AppImage files that provide empty desktop entries (https://github.com/vinifmor/bauh/issues/186)[#186]
 
+### Fixes
+- AppImage
+  - search: not matching manually installed applications
+
+
 ## [0.9.18] 2021-06-18
 ### Fixes
 - Arch